Parabool Jakob feedback fix

+++ 2D_L-Systemen/LSystem2D.h
@@ -15,15 +15,14 @@ class LSystem2D {
 public:
     static img::EasyImage parseLSystem2D(const ini::Configuration &configuration);
 
-    static std::string getString(const LParser::LSystem2D& system);
-
+    static std::string getString(const LParser::LSystem2D &system);
 
 
 private:
 
-    static LParser::LSystem2D getLParser2D(const std::string& inputfile);
+    static LParser::LSystem2D getLParser2D(const std::string &inputfile);
 
-    static Lines2D getLines(LParser::LSystem2D& system2D, std::string basicString, const Color& lineColor);
+    static Lines2D getLines(LParser::LSystem2D &system2D, std::string basicString, const Color &lineColor);
 
     static std::vector<double> scaleLines(Lines2D &lines, int maxSize);
 
@@ -31,14 +30,21 @@ private:
 
 public:
 
-    static void drawLines2D(const Lines2D& lines, img::EasyImage &image);
+    static void drawLines2D(const Lines2D &lines, img::EasyImage &image);
+
     static void drawZBufferLines(img::EasyImage &image, const Lines2D &lines);
+
     static void draw_zbuf_line(ZBuffer &zbuffer, img::EasyImage &image, unsigned int x0,
-                               unsigned int y0, unsigned int z0, unsigned int x1, unsigned int y1, unsigned int z1,
+                               unsigned int y0, double z0, unsigned int x1, unsigned int y1, double z1,
                                const Color &color);
-    static void drawtrifig(img::EasyImage &image, const Figures3d &figures  , double &Xmin,
+
+    static void drawtrifig(img::EasyImage &image, const Figures3d &figures, double &Xmin,
                            double &Xmax, double &Ymin, double &Ymax);
-    static void draw_zbuf_triag(ZBuffer &zBuffer, img::EasyImage &image, Vector3D const& A, Vector3D const& B, Vector3D const& C, double d, double dx, double dy, img::Color color);
+
+    static void
+    draw_zbuf_triag(ZBuffer &zBuffer, img::EasyImage &image, Vector3D const &A, Vector3D const &B, Vector3D const &C,
+                    double d, double dx, double dy, img::Color color);
+
     static void drawZBufferLinesUsingInterpolation(img::EasyImage &image, const Lines2D &lines);
